+++ src/com/intellij/idea/plugin/hybris/impex/codeInsight/lookup/ImpExLookupElementFactory.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019 EPAM Systems <hybrisideaplugin@epam.com>
+ * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -43,7 +43,8 @@ object ImpExLookupElementFactory {
         .withIcon(AllIcons.FileTypes.JavaClass)
 
     fun buildModifierValue(lookupElement: String) = LookupElementBuilder.create(lookupElement)
-    fun buildModifierValue(lookupElement: String, typeText: String) = LookupElementBuilder.create(lookupElement)
+    fun buildModifierValue(lookupElement: String, typeText: String, presentableText: String = lookupElement) = LookupElementBuilder.create(lookupElement)
+        .withPresentableText(presentableText)
         .withTypeText(typeText, true)
 
     fun buildUserRights() = LookupElementBuilder.create(